requirements.txt
requests
mysql-connector
psycopg2-binary